refactor: extract running-config approval level into shared hook
- Move ToolExecutionLevel and normalizeLevel to utils/approval
- Add useAgentRunningConfigApprovalLevel hook that follows selectedAgent
- Replace duplicated fetch logic in Chat and Inbox with the shared hook
- Fixes stale closure concern in Inbox empty-deps useEffect
